Fixing some components and moving some files for better reading
