+++ public/index.html
@@ -5,18 +5,14 @@
     <link rel="icon" href="%PUBLIC_URL%/favicon.ico" />
     <meta name="viewport" content="width=device-width, initial-scale=1" />
     <meta name="theme-color" content="#000000" />
-    <meta
-      name="description"
-      content="LinkTree Vladimir Castañeda"
-    />
+    <meta name="description" content="LinkTree Vladimir Castañeda" />
     <link rel="apple-touch-icon" href="%PUBLIC_URL%/logo192.png" />
     <link rel="manifest" href="%PUBLIC_URL%/manifest.json" />
-   
-    <title>Vladimir Castañeda</title>
+
+    <title>Vladimir Castañeda - Linktree</title>
   </head>
   <body>
     <noscript>You need to enable JavaScript to run this app.</noscript>
     <div id="root"></div>
-  
   </body>
 </html>
