+++ src/App.js
@@ -1,27 +1,13 @@
-import { createGlobalStyle } from "styled-components"
-import NavBar from "./components/Navbar/NavBar"
-import Layout from "./components/Layout/Layout"
-
-const GlobalStyle = createGlobalStyle`
-  * {
-    margin: 0;
-    padding: 0;
-    box-sizing: border-box;
-    font-family: 'Courier New', Courier, monospace;
-  }
-  body {
-    background-color: #E8F6EF;
-    overflow-x: hidden;
-  }
-`
+import { GlobalStyles } from "./GlobalStyles";
+import NavBar from "./components/Navbar/NavBar";
+import Layout from "./components/Layout/Layout";
 
 function App() {
   return (
     <>
-        <GlobalStyle />
-        <NavBar />
-        <Layout />
-        
+      <GlobalStyles />
+      <NavBar />
+      <Layout />
     </>
   );
 }
