+++ src/GlobalStyles/index.js
@@ -1,0 +1,14 @@
+import { createGlobalStyle } from "styled-components";
+
+export const GlobalStyles = createGlobalStyle`
+  * {
+    margin: 0;
+    padding: 0;
+    box-sizing: border-box;
+    font-family: 'Courier New', Courier, monospace;
+  }
+  body {
+    background-color: #E8F6EF;
+    overflow-x: hidden;
+  }
+`;
